+++ aegis/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* [PR-75](https://github.com/AGH-CEAI/aegis_ros/pull/75) - Changed `pymoveit2` upstream repo to a [custom fork](https://github.com/macmacal/pymoveit2).
 * [PR-68](https://github.com/AGH-CEAI/aegis_ros/pull/68) - Updated repos for new robotiq_hande_driver fix.
 
 ### Deprecated